+++ src/controllers/blocks/BlocksController.ts
@@ -1,9 +1,14 @@
 import { ApiPromise } from '@polkadot/api';
+import { isHex } from '@polkadot/util';
 import { RequestHandler } from 'express';
 
 import { BlocksService } from '../../services';
 import { INumberParam } from '../../types/requests';
 import AbstractController from '../AbstractController';
+
+interface ControllerOptions {
+	finalizes: boolean;
+}
 
 /**
  * GET a block.
@@ -38,7 +43,9 @@ import AbstractController from '../AbstractController';
  *   - `method`: Extrinsic method.
  *   - `signature`: Object with `signature` and `signer`, or `null` if unsigned.
  *   - `nonce`: Account nonce, if applicable.
- *   - `args`: Array of arguments.
+ *   - `args`: Array of arguments. Note: if you are expecting an [`OpaqueCall`](https://substrate.dev/rustdocs/v2.0.0/pallet_multisig/type.OpaqueCall.html)
+ * 			and it is not decoded in the response (i.e. it is just a hex string), then Sidecar was not
+ * 			able to decode it and likely that it is not a valid call for the runtime.
  *   - `tip`: Any tip added to the transaction.
  *   - `hash`: The transaction's hash.
  *   - `info`: `RuntimeDispatchInfo` for the transaction. Includes the `partialFee`.
@@ -62,10 +69,8 @@ import AbstractController from '../AbstractController';
  * - `OnInitialize`: https://crates.parity.io/frame_support/traits/trait.OnInitialize.html
  * - `OnFinalize`: https://crates.parity.io/frame_support/traits/trait.OnFinalize.html
  */
-export default class BlocksController extends AbstractController<
-	BlocksService
-> {
-	constructor(api: ApiPromise) {
+export default class BlocksController extends AbstractController<BlocksService> {
+	constructor(api: ApiPromise, private readonly options: ControllerOptions) {
 		super(api, '/blocks', new BlocksService(api));
 		this.initRoutes();
 	}
@@ -88,16 +93,36 @@ export default class BlocksController extends AbstractController<
 		res
 	) => {
 		const eventDocsArg = eventDocs === 'true';
-		const extrsinsicDocsArg = extrinsicDocs === 'true';
+		const extrinsicDocsArg = extrinsicDocs === 'true';
 
-		const hash =
-			finalized === 'false'
-				? (await this.api.rpc.chain.getHeader()).hash
-				: await this.api.rpc.chain.getFinalizedHead();
+		let hash, queryFinalizedHead, omitFinalizedTag;
+
+		// If the network chain doesn't finalize blocks, we dont want a finalized tag.
+		if (!this.options.finalizes) {
+			omitFinalizedTag = true;
+			queryFinalizedHead = false;
+			hash = (await this.api.rpc.chain.getHeader()).hash;
+		} else if (finalized === 'false') {
+			omitFinalizedTag = false;
+			queryFinalizedHead = true;
+			hash = (await this.api.rpc.chain.getHeader()).hash;
+		} else {
+			omitFinalizedTag = false;
+			queryFinalizedHead = false;
+			hash = await this.api.rpc.chain.getFinalizedHead();
+		}
+
+		const options = {
+			eventDocs: eventDocsArg,
+			extrinsicDocs: extrinsicDocsArg,
+			checkFinalized: false,
+			queryFinalizedHead,
+			omitFinalizedTag,
+		};
 
 		BlocksController.sanitizedSend(
 			res,
-			await this.service.fetchBlock(hash, eventDocsArg, extrsinsicDocsArg)
+			await this.service.fetchBlock(hash, options)
 		);
 	};
 
@@ -111,18 +136,28 @@ export default class BlocksController extends AbstractController<
 		{ params: { number }, query: { eventDocs, extrinsicDocs } },
 		res
 	): Promise<void> => {
+		const checkFinalized = isHex(number);
+
 		const hash = await this.getHashForBlock(number);
 
 		const eventDocsArg = eventDocs === 'true';
-		const extrinsinsicDocsArg = extrinsicDocs === 'true';
+		const extrinsicDocsArg = extrinsicDocs === 'true';
 
+		const queryFinalizedHead = !this.options.finalizes ? false : true;
+		const omitFinalizedTag = !this.options.finalizes ? true : false;
+
+		const options = {
+			eventDocs: eventDocsArg,
+			extrinsicDocs: extrinsicDocsArg,
+			checkFinalized,
+			queryFinalizedHead,
+			omitFinalizedTag,
+		};
+
+		// We set the last param to true because we haven't queried the finalizedHead
 		BlocksController.sanitizedSend(
 			res,
-			await this.service.fetchBlock(
-				hash,
-				eventDocsArg,
-				extrinsinsicDocsArg
-			)
+			await this.service.fetchBlock(hash, options)
 		);
 	};
 }